MenuクラスをCLIクラスに統合
